Starting to build out the interface
I started building out the interface. For example, i now have a list of
inputs that correspond to quit, flip and hint. What i have yet to do is
figure out how I am going to deal with counter for the cards. Currently,
it increases with every input, but that is not what I want