Fixing flipped issue
I reported an issue with the cards not being able to flip back and forth.
I added a flipped property to the Card struct and a flip method to control the
flipped state. I also needed to pass in a pointer of the card to one of the template
functions so that it can modify the card